Implemented video upload, hash meta creation and refactored code
This commit introduces video upload functionality and the ability to extract frames from videos. It also includes the generation of blurhash for images and videos and storing them as metadata. The upload and view routes have been refactored for better readability and separation of concerns. Scripts and utilities to support these features have been also added.